Single file implementation, without packaging 🤖
