--- inputClass.java
@@ -1,10 +1,0 @@
-package src.Input;
-// nameOfFirstThingApplicationUsesHereSuchAsInputClass
-public class  inputClass {
-   public inputClass(String value) {
-     tellInformation(value);
-   }
-   public void tellInformation(String value) {
-     System.out.println(value);
-   }
-}
